Modularize Data Structures in nvimrserver
- Extracted data structures into separate file (data_structures.h).
- Included new data_structures.h in nvimrserver.c.
- Updated Makefile to compile the new data_structures.c file as part of nvimrserver.
