+++ R/nvimcom/src/apps/Makefile
@@ -3,8 +3,8 @@ CFLAGS = -pthread -std=gnu99 -O2 -Wall
 
 all: nvimrserver
 
-nvimrserver: nvimrserver.c utilities.c
-	$(CC) $(CFLAGS) nvimrserver.c utilities.c -o nvimrserver
+nvimrserver: nvimrserver.c utilities.c data_structures.c
+	$(CC) $(CFLAGS) nvimrserver.c utilities.c data_structures.c -o nvimrserver
 
 clean:
